requirements.txt
aws-cdk-lib~=2.108.0
constructs>=10.0.0,<11.0.0
pydantic~=2.5.0
data-warehouse-common